Feat: 응답 등록

+++ src/main/java/gdgoc/tuk/official/question/dto/QuestionOrderResponse.java
@@ -1,3 +1,3 @@
 package gdgoc.tuk.official.question.dto;
 
-public record QuestionOrderResponse(Long questionId, Long order) {}
+public record QuestionOrderResponse(Long questionId, Integer order) {}
