+++ src/main/java/gdgoc/tuk/official/question/dto/UpdatedQuestionOrder.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class UpdatedQuestionOrder {
     private Long questionId;
-    private Long order;
+    private Integer order;
 
     public void updateQuestionId(final Long savedQuestionId){
         this.questionId = savedQuestionId;
